+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add cli tests
 - Fix domain query responses
+- Resolve account by starname functionality
 - Normalize fee parameter names
 
 ## v0.4.1